data-server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.8.4
aiohttp-sse-client==0.2.1
aiosignal==1.3.2
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.2.0
cmake==4.0.2
datasets==2.2.0
dill==0.4.0
exceptiongroup==1.3.0
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
hf-xet==1.1.0
huggingface-hub==0.25.2
idna==3.10
InstructorEmbedding==1.0.0
Jinja2==3.1.6
joblib==1.5.0
jsonlines==4.0.0
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psutil==5.9.5
pyarrow==20.0.0
pydantic==1.10.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.26.0
responses==0.18.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.26.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.65.0
transformers==4.28.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.22.0
xxhash==3.5.0
yarl==1.20.0
